Use Convex dev for branch previews
