+++ apps/meseeks/package.json
@@ -8,8 +8,7 @@
 		"url": "https://github.com/igor9silva/meseeks.git"
 	},
 	"scripts": {
-		"dev": "concurrently -r bun:dev:web bun:dev:db",
-		"dev:preview": "bun ./scripts/dev-preview.ts",
+		"dev": "bun ./scripts/dev.ts",
 		"dev:db": "convex dev",
 		"dev:web": "vite dev --port 3000",
 		"build": "vite build",
@@ -20,8 +19,6 @@
 		"lint": "oxfmt --check src convex lib schemas skills scripts tailwind.config.ts tailwind-safelist.ts vite.config.ts postcss.config.cjs components.json convex.json package.json tsconfig.json && oxlint src convex lib schemas skills scripts tailwind.config.ts tailwind-safelist.ts vite.config.ts",
 		"format": "oxfmt src convex lib schemas skills scripts tailwind.config.ts tailwind-safelist.ts vite.config.ts postcss.config.cjs components.json convex.json package.json tsconfig.json",
 		"preview:attach": "bun ./scripts/attach-preview-env.ts",
-		"preview:deploy": "bun ./scripts/deploy-preview.ts",
-		"preview:watch": "bun ./scripts/watch-preview-deploy.ts",
 		"sync-env-to-preview": "bun ./scripts/sync-env-to-preview.ts"
 	},
 	"keywords": [],
